Move rodio dep to frontend
