+++ Cargo.toml
@@ -33,6 +33,7 @@ bytesize = "1.3.0"
 timeago = "0.4.2"
 futures = "0.3.30"
 itertools = "0.13.0"
+rodio = "0.19.0"
 
 [dependencies.uuid]
 version = "1.0"
