+++ data/Cargo.toml
@@ -34,7 +34,6 @@ timeago = "0.4.2"
 url = { version = "2.5.0", features = ["serde"] }
 regex = "1.10.4"
 walkdir = "2.5.0"
-rodio = "0.19.0"
 once_cell = "1.19.0"
 
 [dependencies.irc]
